Move SVG related code to SvgImageDescriptor
