+++ bundles/org.eclipse.jface/pom.xml
@@ -20,7 +20,7 @@
   </parent>
   <groupId>org.eclipse.jface</groupId>
   <artifactId>org.eclipse.jface</artifactId>
-  <version>3.15.100.lgc202002271800</version>
+  <version>3.15.100.lgc202006261800</version>
   <packaging>eclipse-plugin</packaging>
 
   <properties>
